lisätty liigoihin ominaisuus pitää ne avoimena tai suljettuna ja lisätty liigalle salasana ominaisuus

+++ backend/src/controllers/league.js
@@ -1,6 +1,7 @@
 const leagueRouter = require('express').Router()
 const League = require('../models/leagues')
 const middleware = require('../utils/middleware')
+const bcrypt = require('bcrypt')
 
 leagueRouter.get('/', async (request, response) => {
   const leagues = await League
@@ -8,14 +9,29 @@ leagueRouter.get('/', async (request, response) => {
   response.json(leagues)
 })
 
-// Uuden kimpan luonti
+// Uuden liigan luonti
 leagueRouter.post('/', middleware.userExtractor, async (request, response) => {
   const body = request.body
+  const title = body.title
 
   const user = request.user
+
+  const existingLeague = await League.findOne({ title })
+  if (existingLeague) {
+    return response.status(400).json({
+      error: 'league name must be unique'
+    })
+  }
   if (!user) {
     return response.status(401).json({ error: 'token missing or invalid' })
   }
+  const password = body.password
+
+  if (!password) {
+    return response.status(401).json({ error: 'password missing' })
+  }
+  const saltRounds = 10
+  const passwordHash = await bcrypt.hash(password, saltRounds)
 
 
 
@@ -23,6 +39,9 @@ leagueRouter.post('/', middleware.userExtractor, async (request, response) => {
     title: body.title,
     playerCount: 1,
     creator: user._id,
+    status: body.status,
+    passwordHash: passwordHash
+
   })
   const savedLeague = await league.save()
   user.leagues = user.leagues.concat(savedLeague._id)
@@ -37,11 +56,54 @@ leagueRouter.delete('/:id', middleware.userExtractor, async (request, response) 
   if (!user) {
     return response.status(401).json({ error: 'token missing or invalid' })
   }
+
+  if (!league) {
+    return response.status(401).json('League has already been deleted')
+  }
   if (league.creator.toString() === user._id.toString()) {
     await League.findByIdAndDelete(request.params.id)
-    response.status(204).end()
+    return response.status(201).json('League deleted')
   } else {
-    return response.status(401).end()
+    return response.status(401).json('Only league creator can delete league')
+  }
+
+
+
+})
+//Liigan luoja pystyy muokkaamaan liigan avoimuutta
+leagueRouter.post('/:id', middleware.userExtractor, async (request, response) => {
+  const league = await League.findById(request.params.id)
+  const user = request.user
+  const body = request.body
+  if (!user) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+  if (league.creator.toString() === user._id.toString()) {
+
+    const password = body.password
+
+    if (!password) {
+      return response.status(401).json({ error: 'password missing' })
+    }
+
+    const saltRounds = 10
+    const passwordHash = await bcrypt.hash(password, saltRounds)
+
+
+
+    const update = {
+      title: league.title,
+      creator: league.creator,
+      status: body.status,
+      passwordHash: passwordHash,
+      users: league.users
+
+
+    }
+    const updatedLeague = await League.findByIdAndUpdate(request.params.id, update)
+    response.status(201).json(updatedLeague)
+  } else {
+    return response.status(401).json('Only league creator can delete league')
   }
 
 
@@ -54,7 +116,18 @@ leagueRouter.put('/:id', middleware.userExtractor, async (request, response) => 
   if (!user) {
     return response.status(401).json({ error: 'token missing or invalid' })
   }
-  console.log(league.users)
+  const password = request.body.password
+  if (league.status === "closed") {
+    const passwordCorrect = user === null
+      ? false
+      : await bcrypt.compare(password, league.passwordHash)
+
+    if (!(user && passwordCorrect)) {
+      return response.status(401).json({
+        error: 'invalid password'
+      })
+    }
+  }
   const userAlreadyJoined = league.users.find((element) => element == user._id.toString())
   if (userAlreadyJoined) {
     return response.status(400).json({
